* аdded israid() DBC check. * аdded Transporter information is now handled in Unit.cpp, Vehicle code has been changed and Changing seats has been disabled for a while. * modified Rewrote the pet int thing in Pet.cpp from the previous revision, this removes unsigned signed warning. * modified Fixed crash in Creature Spawn where client is saying we are on a transport, it now checks if transport exists. * modified .debug SetPhase 0 will now reset players phase. * modified Difficulty achievements now require the correct difficulty, maybe we can do this a different way, but this works for now.

+++ src/WorldServer/Object.h
@@ -547,7 +547,7 @@ protected:
 	// Mark values that player should get when he/she/it sees object for first time.
 	virtual void _SetCreateBits(UpdateMask *updateMask, Player* target) const;
 
-	void _BuildMovementUpdate( ByteBuffer *data, uint16 flags, uint32 flags2, Player* target );
+	void _BuildMovementUpdate( ByteBuffer * data, uint32 flags, uint32 flags2, Player* target );
 	void _BuildValuesUpdate( ByteBuffer *data, UpdateMask *updateMask, Player* target );
 
 	/* Main Function called by isInFront(); */
